In the beginning, Davis created the project.
And the project was without form, and void; and unwritten features
were upon the face of the deep. And Davis said, "Let there be a
Github standardized Scala gitignore file". And Davis saw the gitignore
file, and it was good. And Davis said, "Let there be a LICENSE file
containing a 3-BSD Revised License". And Davis saw the LICENSE file,
and it was good. And Davis said, "Let there be some default information
about this project". And Davis called this information a README.

And the gitignore and the LICENSE and the README were the first commit.