Added BufferReader to Buffer types.
BufferReader is currently backed by a Nett3y ChannelBuffer, and is
a near replica of the existing code for finagle-mysql, with the
exception that the ordering is Big Endian in conformance with
Postgres standard.

It is expected that superfulous methods will be removed from the
BufferReader / BufferWriter upon completed of Packet encoding
and decoding.