Prepare repository for turning public on Github.
Now that we can parse even small amounts of data from a Result,
it's time to turn Repository public. This includes a detailed README,
hooking up code coverage via CodeCov, and placing the open source
badges on the README as well.

The build and errata was modified in this commit, no working code was
changed.