Completed Simple Query state machine message flow.
The Simple Query messgage flow described here:
http://www.postgresql.org/docs/current/static/protocol-flow.html#AEN108664
has been completed. We can decode all relevant Packets, though
ErrorResponses and NoticeResponses are throwing Exceptions until we can
classify them further. We are also making a best effort to inform a
Postgresql Server process that we are terminating whenever we close.

The Result type is a bit of a mess, but cleaning that up requires
further knowledge.