Bedazzled Column type, removed extraneous types.
The Column type was cleaned up, Scaladoc information was added for both
Column and FormatCode.

PostgresqlType was removed as it is beyond the scope of core - client
code, or another module, will handle the conversion of a data into
a specific type. This addresses
https://github.com/penland365/roc/issues/26 .

Issues with test data generation relating to DataRow were also fixed,
see https://github.com/penland365/roc/issues/20 .