Introduced the Buf type as Readable / Writeable Array Buffer.
The ChannelBuffer backed implementation of Buffer used currently ( and
borrowed from finagle-mysql ) has 3 issues - it generates a lot of
garbage with separate Reader / Writer classes, it's based on Netty3, and
it doesn't correctly write all UTF-8 encoded strings to it's buffer.

The Buf type was introduced to address this. As of this commit, nothing
internal to roc-core will use this new type outside of tests. Conversion
of existing Buffer uses will take place gradually prior to finagle-4
being introduced.

There is still work to be done in completing Benchmarks.

This partially addresses https://github.com/finagle/roc/issues/48