Added Boolean Element Decoder.
An Element Decoder for Booleans was added to the types project. This
partiall addresses https://github.com/finagle/roc/issues/51 . As a
result, the generators test object for types was no longer needed, so it
was removed.

In addition, a bug was fixed where we were eating Errors thrown during
Query transmission. The Decoded error is now re-thrown by
Future.exception.